// employee-managementUI/src/app/models/employeeModel.ts
export interface Employee{
    id: number;
    employeeId: number;
    firstName: string;
    lastName: string;
    birthDate: Date;
    nid: string;
    username: string;
    passwordHash: string;
    email: string;
    phoneNumber: string;
    address: string;
    city: string;
    district: string;
    division: string;
    country: string;
    role: string;
    department: string;
    position: string;
    hireDate: Date;
    salary: number;
    photoUrl: string;
    nidUrl: string;
}